[#174911379] testing engine file processor
